Add repository query methods to fetch specified collections of data
- Add methods to GameSummaryRepository and QuizRepository for fetching different collections of data
- Add repository test classes with tests covering basics

+++ src/main/java/se/juninatt/quizwiz/model/entity/GameSummary.java
@@ -23,7 +23,7 @@ public class GameSummary {
     public GameSummary() {}
 
     public GameSummary(int totalScore,
-                       int timeUsedPercentage,
+                       double timeUsedPercentage,
                        double completionPercentage,
                        String date) {
         this.totalScore = totalScore;
